requirements.txt
streamlit>=1.35
pandas>=2.2
numpy>=1.26
pyyaml>=6.0
bcrypt>=4.1
streamlit-authenticator>=0.3
python-docx>=1.1
python-pptx>=0.6
openpyxl>=3.1
Pillow>=10.0
matplotlib>=3.8